geo_api_django/requirements.txt
asgiref==3.11.0
Django==6.0
django-environ==0.12.0
django-rest-authtoken==2.1.4
djangorestframework==3.16.1
djangorestframework-jwt==1.11.0
dotenv==0.9.9
iniconfig==2.3.0
packaging==25.0
pluggy==1.6.0
psycopg2-binary==2.9.11
Pygments==2.19.2
PyJWT==1.7.1
pytest==9.0.2
pytest-django==4.11.1
python-dotenv==1.2.1
sqlparse==0.5.5
